Restructure non-model files into training and core directories
