+++ larq_zoo/__init__.py
@@ -1,7 +1,6 @@
 import larq_zoo.literature as literature
 import larq_zoo.sota as sota
 from larq_zoo import datasets
-from larq_zoo.data import preprocess_input
-from larq_zoo.utils import decode_predictions
+from larq_zoo.core.utils import decode_predictions
 
 __all__ = ["datasets", "decode_predictions", "preprocess_input", "literature", "sota"]
