+++ larq_zoo/sota/quicknet.py
@@ -4,8 +4,8 @@ import larq as lq
 import tensorflow as tf
 from zookeeper import Field, factory
 
-from larq_zoo import utils
-from larq_zoo.model_factory import ModelFactory
+from larq_zoo.core import utils
+from larq_zoo.core.model_factory import ModelFactory
 
 
 def LCEFirstLayer(filters: int, x: tf.Tensor) -> tf.Tensor:
